Major refactor - WindowState struct
This commit introduces a major refactor with the addition of a
WindowState C-style struct that encapsulates the current window
and its properties, including the text of the file currently displayed.

This introduces numerous benefits. Using a standard logic for storing
and referencing the window and its properties removes a lot of the borrow
headaches I was wasting time fighting.

Additionally, window properties no longer need to be tracked as mutable
variables within the main function, which allows for logic to ported out
of the main function into seperate functions. This means that a generic
jump_to_line function can be used to fulfill controls such as up/down a
line, as well as new controlls like up/down half screen and jump to top.