Add search functionality
This commit adds search functionality, both forward and reverse, as well
as the ability to jump forward and backward through search results.

Architecturally, this starts with a search scraper function that
returns a vector of tuples corresponding to line and character numbers
of occurences of a search term. This is called from a WindowState
function which processes the user's search input, feeds it to the
scraper, updates the state with the search results, and jumps to the
first result. Additionally, there are jump to next/last search result
functions, which use jump_to_line to move to the search result and then
rotate the list of results.